got the controller to return a response

+++ backend/src/main/java/com/job_intel/backend/Dtos/comboDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class comboDTO {
 
     private String[] combo;
-    private int rankNum;
+    private Long rankNum;
 }
